refatored vscodeUriHandler to expose connection view helper functions
